Refactor download generation and reduce bulk download modal to one step
