+++ app/assets/src/components/views/bulk_download/BulkDownloadModal.jsx
@@ -1,8 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { unset, find, get, set } from "lodash/fp";
-import memoize from "memoize-one";
 
+import {
+  getBackgrounds,
+  uploadedByCurrentUser,
+  getHeatmapMetrics,
+} from "~/api";
 import {
   getBulkDownloadTypes,
   validateSampleIdsForBulkDownload,
@@ -10,39 +13,6 @@ import {
 import Modal from "~ui/containers/Modal";
 
 import ChooseStep from "./ChooseStep";
-import ReviewStep from "./ReviewStep";
-
-const assembleSelectedDownload = memoize(
-  (
-    selectedDownloadTypeName,
-    allSelectedFields,
-    allSelectedFieldsDisplay,
-    sampleIds
-  ) => {
-    const fieldValues = get(selectedDownloadTypeName, allSelectedFields);
-    const fieldDisplayNames = get(
-      selectedDownloadTypeName,
-      allSelectedFieldsDisplay
-    );
-
-    const fields = {};
-    if (fieldValues) {
-      for (let [fieldName, fieldValue] of Object.entries(fieldValues)) {
-        fields[fieldName] = {
-          value: fieldValue,
-          // Use the display name for the value if it exists. Otherwise, use the value.
-          displayName: fieldDisplayNames[fieldName] || fieldValue,
-        };
-      }
-    }
-
-    return {
-      downloadType: selectedDownloadTypeName,
-      fields,
-      sampleIds: Array.from(sampleIds),
-    };
-  }
-);
 
 class BulkDownloadModal extends React.Component {
   state = {
@@ -58,38 +28,44 @@ class BulkDownloadModal extends React.Component {
     // in the params. This is also how the bulk download is stored in the database.
     selectedFieldsDisplay: {},
     selectedDownloadTypeName: null,
-    currentStep: "choose",
     validSampleIds: new Set(),
     invalidSampleNames: [],
     validationError: null,
+    backgroundOptions: null,
+    metricsOptions: null,
+    allSamplesUploadedByCurrentUser: false,
+    loading: true,
   };
 
   componentDidMount() {
-    this.fetchDownloadTypesAndValidationInfo();
+    this.parallelFetchNeededData();
   }
 
-  componentDidUpdate(prevProps) {
-    // If the user has just closed the modal, reset it.
-    if (prevProps.open && !this.props.open) {
-      this.setState({
-        selectedDownloadTypeName: null,
-        currentStep: "choose",
-        selectedFields: {},
-      });
-    }
-  }
+  // *** Async requests ***
 
-  async fetchDownloadTypesAndValidationInfo() {
+  async parallelFetchNeededData() {
     const { selectedSampleIds } = this.props;
 
     const bulkDownloadTypesRequest = this.fetchDownloadTypes();
     const sampleValidationInfoRequest = this.fetchValidationInfo(
       Array.from(selectedSampleIds)
     );
+    const backgroundOptionsRequest = this.fetchBackgrounds();
+    const metricsOptionsRequest = this.fetchHeatmapMetrics();
+    const allSamplesUploadedByCurrentUserRequest = this.checkAllSamplesUploadedByCurrentUser();
 
-    const [bulkDownloadTypes, sampleValidationInfo] = await Promise.all([
+    const [
+      bulkDownloadTypes,
+      sampleValidationInfo,
+      backgroundOptions,
+      metricsOptions,
+      allSamplesUploadedByCurrentUser,
+    ] = await Promise.all([
       bulkDownloadTypesRequest,
       sampleValidationInfoRequest,
+      backgroundOptionsRequest,
+      metricsOptionsRequest,
+      allSamplesUploadedByCurrentUserRequest,
     ]);
 
     const validSampleIds = new Set(sampleValidationInfo.validSampleIds);
@@ -101,6 +77,10 @@ class BulkDownloadModal extends React.Component {
       validSampleIds,
       invalidSampleNames,
       validationError,
+      backgroundOptions,
+      metricsOptions,
+      allSamplesUploadedByCurrentUser,
+      loading: false,
     });
   }
 
@@ -118,102 +98,64 @@ class BulkDownloadModal extends React.Component {
     return sampleValidationInfo;
   }
 
-  handleSelectDownloadType = selectedDownloadTypeName => {
-    this.setState({
-      selectedDownloadTypeName,
-    });
-  };
+  // TODO(mark): Set a reasonable default background based on the samples and the user's preferences.
+  async fetchBackgrounds() {
+    const backgrounds = await getBackgrounds();
 
-  handleFieldSelect = (downloadType, fieldType, value, displayName) => {
-    this.setState(prevState => {
-      // If the value is undefined, delete it from selectedFields.
-      // This allows us to support cases where certain fields are conditionally required;
-      // if the field becomes no longer required, we can unset it.
-      const newSelectedFields =
-        value !== undefined
-          ? set([downloadType, fieldType], value, prevState.selectedFields)
-          : unset([downloadType, fieldType], prevState.selectedFields);
+    const backgroundOptions = backgrounds.map(background => ({
+      text: background.name,
+      value: background.id,
+    }));
 
-      const newSelectedFieldsDisplay =
-        displayName !== undefined
-          ? set(
-              [downloadType, fieldType],
-              displayName,
-              prevState.selectedFieldsDisplay
-            )
-          : unset([downloadType, fieldType], prevState.selectedFieldsDisplay);
+    return backgroundOptions;
+  }
 
-      return {
-        selectedFields: newSelectedFields,
-        selectedFieldsDisplay: newSelectedFieldsDisplay,
-      };
-    });
-  };
+  // We use the heatmap metrics as the valid metrics for bulk downloads.
+  async fetchHeatmapMetrics() {
+    const heatmapMetrics = await getHeatmapMetrics();
 
-  handleChooseStepContinue = () => {
-    this.setState({ currentStep: "review" });
-  };
-  handleBackClick = () => {
-    this.setState({ currentStep: "choose" });
-  };
+    return heatmapMetrics;
+  }
 
-  renderStep = () => {
+  async checkAllSamplesUploadedByCurrentUser() {
+    const { selectedSampleIds } = this.props;
+    const allSamplesUploadedByCurrentUser = await uploadedByCurrentUser(
+      Array.from(selectedSampleIds)
+    );
+
+    return allSamplesUploadedByCurrentUser;
+  }
+
+  // *** Render methods ***
+
+  render() {
+    const { open, onGenerate } = this.props;
     const {
-      currentStep,
       bulkDownloadTypes,
-      selectedDownloadTypeName,
-      selectedFields,
-      selectedFieldsDisplay,
       validSampleIds,
       invalidSampleNames,
       validationError,
+      backgroundOptions,
+      metricsOptions,
+      allSamplesUploadedByCurrentUser,
+      loading,
     } = this.state;
-
-    if (currentStep === "choose") {
-      return (
-        <ChooseStep
-          downloadTypes={bulkDownloadTypes}
-          selectedDownloadTypeName={selectedDownloadTypeName}
-          onSelect={this.handleSelectDownloadType}
-          selectedFields={get(selectedDownloadTypeName, selectedFields)}
-          onFieldSelect={this.handleFieldSelect}
-          onContinue={this.handleChooseStepContinue}
-          validSampleIds={validSampleIds}
-          invalidSampleNames={invalidSampleNames}
-          validationError={validationError}
-        />
-      );
-    }
-
-    if (currentStep === "review") {
-      const selectedDownload = assembleSelectedDownload(
-        selectedDownloadTypeName,
-        selectedFields,
-        selectedFieldsDisplay,
-        validSampleIds
-      );
-
-      const selectedDownloadType = find(
-        ["type", selectedDownloadTypeName],
-        bulkDownloadTypes
-      );
-
-      return (
-        <ReviewStep
-          selectedDownload={selectedDownload}
-          downloadType={selectedDownloadType}
-          onBackClick={this.handleBackClick}
-        />
-      );
-    }
-  };
-
-  render() {
-    const { open } = this.props;
 
     return (
       <Modal narrow open={open} tall onClose={this.props.onClose}>
-        {this.renderStep()}
+        {loading && "L O A D I N G"}
+        {!loading && (
+          <ChooseStep
+            downloadTypes={bulkDownloadTypes}
+            onGenerate={onGenerate}
+            validSampleIds={validSampleIds}
+            invalidSampleNames={invalidSampleNames}
+            validationError={validationError}
+            backgroundOptions={backgroundOptions}
+            metricsOptions={metricsOptions}
+            allSamplesUploadedByCurrentUser={allSamplesUploadedByCurrentUser}
+          />
+        )}
       </Modal>
     );
   }
@@ -223,6 +165,8 @@ BulkDownloadModal.propTypes = {
   onClose: PropTypes.func.isRequired,
   open: PropTypes.bool,
   selectedSampleIds: PropTypes.instanceOf(Set),
+  // called when a bulk download has successfully been kicked off
+  onGenerate: PropTypes.func.isRequired,
 };
 
 export default BulkDownloadModal;
